Split forecast engine/fit/tests to stay under the file-size cap
The three accuracy fixes pushed sustain-forecast.js, sustain-forecast-fit.js
and sustain-forecast.test.js over the 600/1200-line hard caps. Behaviour-
preserving extraction, no logic change (backtest output bit-identical):

- sustain-forecast-fit-base.js — shared constants, sanity bounds and pure
  helpers (helsinkiHour, labelModes, slopeThruOrigin), required by both
  the fit module and the notes module.
- sustain-forecast-notes.js — operator-facing forecast prose (buildNotes).
- tests/sustain-forecast-tuning.test.js — the destratification, heating-
  thermostat and cloud-reference regression tests.
